+++ packages/node/src/indexer/blockDispatcher/block-dispatcher.service.ts
@@ -1,26 +1,22 @@
-// Copyright 2020-2023 SubQuery Pte Ltd authors & contributors
+// Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
 import { Inject, Injectable, OnApplicationShutdown } from '@nestjs/common';
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import {
   NodeConfig,
-  SmartBatchService,
   StoreCacheService,
   StoreService,
   IProjectService,
-  PoiService,
   PoiSyncService,
   BlockDispatcher,
   ProcessBlockResponse,
   IProjectUpgradeService,
+  IBlock,
 } from '@subql/node-core';
-import {
-  CosmosProjectDs,
-  SubqueryProject,
-} from '../../configure/SubqueryProject';
+import { CosmosDatasource } from '@subql/types-cosmos';
+import { SubqueryProject } from '../../configure/SubqueryProject';
 import { ApiService } from '../api.service';
-import { DynamicDsService } from '../dynamic-ds.service';
 import { IndexerManager } from '../indexer.manager';
 import { BlockContent } from '../types';
 
@@ -29,46 +25,37 @@ import { BlockContent } from '../types';
  */
 @Injectable()
 export class BlockDispatcherService
-  extends BlockDispatcher<BlockContent, CosmosProjectDs>
+  extends BlockDispatcher<BlockContent, CosmosDatasource>
   implements OnApplicationShutdown
 {
   constructor(
-    private apiService: ApiService,
+    apiService: ApiService,
     nodeConfig: NodeConfig,
     private indexerManager: IndexerManager,
     eventEmitter: EventEmitter2,
-    @Inject('IProjectService') projectService: IProjectService<CosmosProjectDs>,
+    @Inject('IProjectService')
+    projectService: IProjectService<CosmosDatasource>,
     @Inject('IProjectUpgradeService')
     projectUpgradeService: IProjectUpgradeService,
-    smartBatchService: SmartBatchService,
     storeService: StoreService,
     storeCacheService: StoreCacheService,
     poiSyncService: PoiSyncService,
     @Inject('ISubqueryProject') project: SubqueryProject,
-    dynamicDsService: DynamicDsService,
   ) {
     super(
       nodeConfig,
       eventEmitter,
       projectService,
       projectUpgradeService,
-      smartBatchService,
       storeService,
       storeCacheService,
       poiSyncService,
       project,
-      dynamicDsService,
-      async (blockNums: number[]): Promise<BlockContent[]> => {
-        // If specVersion not changed, a known overallSpecVer will be pass in
-        // Otherwise use api to fetch runtimes
-        return this.apiService.fetchBlocks(blockNums);
-      },
+      apiService.fetchBlocks.bind(apiService),
     );
   }
 
-  async init(
-    onDynamicDsCreated: (height: number) => Promise<void>,
-  ): Promise<void> {
+  async init(onDynamicDsCreated: (height: number) => void): Promise<void> {
     await super.init(onDynamicDsCreated);
   }
 
@@ -77,11 +64,11 @@ export class BlockDispatcherService
   }
 
   protected async indexBlock(
-    block: BlockContent,
+    block: IBlock<BlockContent>,
   ): Promise<ProcessBlockResponse> {
     return this.indexerManager.indexBlock(
       block,
-      await this.projectService.getDataSources(this.getBlockHeight(block)),
+      await this.projectService.getDataSources(block.getHeader().blockHeight),
     );
   }
 }