+++ packages/node/src/indexer/cosmosClient.connection.ts
@@ -1,16 +1,24 @@
-// Copyright 2020-2023 SubQuery Pte Ltd authors & contributors
+// Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
 import { Registry } from '@cosmjs/proto-signing';
 import { HttpEndpoint } from '@cosmjs/stargate';
-import { Tendermint37Client } from '@cosmjs/tendermint-rpc';
 import {
+  CometClient,
+  Tendermint37Client,
+  Comet38Client,
+  Tendermint34Client,
+} from '@cosmjs/tendermint-rpc';
+import {
+  IBlock,
   ApiConnectionError,
   ApiErrorType,
   IApiConnectionSpecific,
   NetworkMetadataPayload,
+  exitWithError,
 } from '@subql/node-core';
 import { getLogger } from '@subql/node-core/dist';
+import { IEndpointConfig } from '@subql/types-core';
 import { CosmosClient, CosmosSafeClient } from './api.service';
 import { HttpClient, WebsocketClient } from './rpc-clients';
 import { BlockContent } from './types';
@@ -18,32 +26,60 @@ import { BlockContent } from './types';
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const { version: packageVersion } = require('../../package.json');
 
-const RETRY_DELAY = 2_500;
+const logger = getLogger('CosmosClientConnection');
 
-const logger = getLogger('cosmos-client-connection');
+/**
+ * Auto-detects the version of the backend and uses a suitable client.
+ */
+async function connectComet(
+  client: WebsocketClient | HttpClient,
+): Promise<CometClient> {
+  // Tendermint/CometBFT 0.34/0.37/0.38 auto-detection. Starting with 0.37 we seem to get reliable versions again 🎉
+  // Using 0.34 as the fallback.
+  let out: CometClient;
+  const tm37Client = await Tendermint37Client.create(client);
+  const version = (await tm37Client.status()).nodeInfo.version;
+  if (version.startsWith('0.37.')) {
+    logger.debug(`Using Tendermint 37 Client`);
+    out = tm37Client;
+  } else if (version.startsWith('0.38.')) {
+    tm37Client.disconnect();
+    logger.debug(`Using Comet 38 Client`);
+    out = await Comet38Client.create(client);
+  } else {
+    tm37Client.disconnect();
+    logger.debug(`Using Tendermint 34 Client`);
+    out = await Tendermint34Client.create(client);
+  }
+  return out;
+}
 
 type FetchFunc = (
   api: CosmosClient,
   batch: number[],
-) => Promise<BlockContent[]>;
+) => Promise<IBlock<BlockContent>[]>;
 
 export class CosmosClientConnection
   implements
-    IApiConnectionSpecific<CosmosClient, CosmosSafeClient, BlockContent[]>
+    IApiConnectionSpecific<
+      CosmosClient,
+      CosmosSafeClient,
+      IBlock<BlockContent>[]
+    >
 {
-  private tmClient: Tendermint37Client;
-  private registry: Registry;
   readonly networkMeta: NetworkMetadataPayload;
 
-  constructor(
+  private constructor(
     public unsafeApi: CosmosClient,
     private fetchBlocksBatches: FetchFunc,
     chainId: string,
+    private cometClient: CometClient,
+    private registry: Registry,
   ) {
     this.networkMeta = {
       chain: chainId,
-      specName: undefined,
-      genesisHash: undefined,
+      specName: undefined as any,
+      genesisHash: undefined as any, // Cant always get the genesis hash because of pruning
     };
   }
 
@@ -51,33 +87,37 @@ export class CosmosClientConnection
     endpoint: string,
     fetchBlocksBatches: FetchFunc,
     registry: Registry,
+    config: IEndpointConfig,
   ): Promise<CosmosClientConnection> {
     const httpEndpoint: HttpEndpoint = {
       url: endpoint,
       headers: {
         'User-Agent': `SubQuery-Node ${packageVersion}`,
+        ...config.headers,
       },
     };
 
     const rpcClient =
       endpoint.includes('ws://') || endpoint.includes('wss://')
         ? new WebsocketClient(endpoint, (err) => {
-            logger.error(err, `Websocket connection failed`);
-            process.exit(1);
+            exitWithError(
+              new Error(`Websocket connection failed`, { cause: err }),
+              logger,
+            );
           })
         : new HttpClient(httpEndpoint);
 
-    const tendermint = await Tendermint37Client.create(rpcClient);
+    const cometClient = await connectComet(rpcClient);
 
-    const api = new CosmosClient(tendermint, registry);
+    const api = new CosmosClient(cometClient, registry);
 
     const connection = new CosmosClientConnection(
       api,
       fetchBlocksBatches,
       await api.getChainId(),
+      cometClient,
+      registry,
     );
-    connection.setTmClient(tendermint);
-    connection.setRegistry(registry);
 
     logger.info(`connected to ${endpoint}`);
 
@@ -85,20 +125,12 @@ export class CosmosClientConnection
   }
 
   safeApi(height: number): CosmosSafeClient {
-    return new CosmosSafeClient(this.tmClient, height);
-  }
-
-  private setTmClient(tmClient: Tendermint37Client): void {
-    this.tmClient = tmClient;
-  }
-
-  private setRegistry(registry: Registry): void {
-    this.registry = registry;
+    return new CosmosSafeClient(this.cometClient, height);
   }
 
   // eslint-disable-next-line @typescript-eslint/require-await
   async apiConnect(): Promise<void> {
-    this.unsafeApi = new CosmosClient(this.tmClient, this.registry);
+    this.unsafeApi = new CosmosClient(this.cometClient, this.registry);
   }
 
   // eslint-disable-next-line @typescript-eslint/require-await
@@ -106,7 +138,7 @@ export class CosmosClientConnection
     this.unsafeApi.disconnect();
   }
 
-  async fetchBlocks(heights: number[]): Promise<BlockContent[]> {
+  async fetchBlocks(heights: number[]): Promise<IBlock<BlockContent>[]> {
     const blocks = await this.fetchBlocksBatches(this.unsafeApi, heights);
     return blocks;
   }