+++ packages/node/src/indexer/worker/worker.module.ts
@@ -1,10 +1,10 @@
-// Copyright 2020-2023 SubQuery Pte Ltd authors & contributors
+// Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
 import { Module } from '@nestjs/common';
 import { EventEmitterModule } from '@nestjs/event-emitter';
 import { ScheduleModule } from '@nestjs/schedule';
-import { DbModule } from '@subql/node-core';
+import { DbModule, WorkerCoreModule } from '@subql/node-core';
 import { ConfigureModule } from '../../configure/configure.module';
 import { WorkerFetchModule } from './worker-fetch.module';
 
@@ -14,6 +14,7 @@ import { WorkerFetchModule } from './worker-fetch.module';
     EventEmitterModule.forRoot(),
     ConfigureModule.register(),
     ScheduleModule.forRoot(),
+    WorkerCoreModule,
     WorkerFetchModule,
   ],
   controllers: [],