+++ packages/node/src/init.ts
@@ -1,11 +1,15 @@
-// Copyright 2020-2023 SubQuery Pte Ltd authors & contributors
+// Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
 import { NestFactory } from '@nestjs/core';
-import { findAvailablePort, notifyUpdates } from '@subql/common';
-import { getLogger, NestLogger } from '@subql/node-core';
+import { notifyUpdates } from '@subql/common';
+import {
+  exitWithError,
+  getLogger,
+  getValidPort,
+  NestLogger,
+} from '@subql/node-core';
 import { AppModule } from './app.module';
-import { ApiService } from './indexer/api.service';
 import { FetchService } from './indexer/fetch.service';
 import { ProjectService } from './indexer/project.service';
 import { yargsOptions } from './yargs';
@@ -13,7 +17,6 @@ const pjson = require('../package.json');
 
 const { argv } = yargsOptions;
 
-const DEFAULT_PORT = 3000;
 const logger = getLogger('subql-node');
 
 notifyUpdates(pjson, logger);
@@ -21,20 +24,7 @@ notifyUpdates(pjson, logger);
 export async function bootstrap(): Promise<void> {
   logger.info(`Current ${pjson.name} version is ${pjson.version}`);
 
-  const validate = (x: any) => {
-    const p = parseInt(x);
-    return isNaN(p) ? null : p;
-  };
-
-  const port = validate(argv.port) ?? (await findAvailablePort(DEFAULT_PORT));
-  if (!port) {
-    logger.error(
-      `Unable to find available port (tried ports in range (${port}..${
-        port + 10
-      })). Try setting a free port manually by setting the --port flag`,
-    );
-    process.exit(1);
-  }
+  const port = await getValidPort(argv.port);
 
   if (argv.unsafe) {
     logger.warn(
@@ -50,10 +40,8 @@ export async function bootstrap(): Promise<void> {
 
     const projectService: ProjectService = app.get('IProjectService');
     const fetchService = app.get(FetchService);
-    const apiService = app.get(ApiService);
 
     // Initialise async services, we do this here rather than in factories, so we can capture one off events
-    await apiService.init();
     await projectService.init();
     await fetchService.init(projectService.startHeight);
 
@@ -63,7 +51,6 @@ export async function bootstrap(): Promise<void> {
 
     logger.info(`Node started on port: ${port}`);
   } catch (e) {
-    logger.error(e, 'Node failed to start');
-    process.exit(1);
+    exitWithError(new Error(`Node failed to start`, { cause: e }), logger);
   }
 }